+++ R/affirm_clean_join.R
@@ -1,12 +1,10 @@
-#' Affirm Class
+#' Affirm Clean Join
 #'
 #' A wrapper for `affirm_true()`.
-#' Reports `columns` that do not inherit `class`, e.g.
-#' `select(data, all_of(columns) && where(\(x) !inherits(x, class)))`
+#' Reports `columns` whose names end with `".x"` or `".y"`, indicating
+#' a sloppy merge.
 #'
 #' @inheritParams affirm_true
-#' @param columns columns to check class
-#' @param class character class to affirm
 #' @param id,priority,data_frames Optional additional information that will be passed to affirmation report.
 #' - `id` must be an integer, e.g. `id = 1L`
 #' - `priority` must be an integer, e.g. `priority = 1L`
@@ -20,50 +18,47 @@
 #' @examples
 #' affirm_init(replace = TRUE)
 #'
-#' affirm_class(
-#'   as_tibble(iris),
-#'   label = "all cols are numeric (but Species really isn't)",
-#'   columns = everything(),
-#'   class = "numeric"
+#' df <-
+#'   dplyr::tibble(lgl = c(NA, TRUE, NA, FALSE, NA)) |>
+#'   dplyr::mutate(id = dplyr::row_number())
+#'
+#' affirm_clean_join(
+#'   dplyr::full_join(df, df, by = "id"),
+#'   label = "Checking for clean merge"
 #' )
 #'
 #' affirm_close()
-affirm_class <- function(data,
-                         label,
-                         columns,
-                         class,
-                         id = NA_integer_,
-                         priority = NA_integer_,
-                         data_frames = NA_character_,
-                         report_listing = NULL,
-                         data_action = NULL,
-                         error = getOption("affirm.error", default = FALSE)) {
+affirm_clean_join <- function(data,
+                              label,
+                              id = NA_integer_,
+                              priority = NA_integer_,
+                              data_frames = NA_character_,
+                              report_listing = NULL,
+                              data_action = NULL,
+                              error = getOption("affirm.error", default = FALSE)) {
   # check and process inputs ---------------------------------------------------
-  if (missing(data) || missing(columns) || missing(class) || missing(label)) {
-    cli::cli_abort("Arguments {.code data}, {.code label}, {.code columns}, and {.code class} are required.")
+  if (missing(data)) {
+    cli::cli_abort("Arguments {.code data} and {.code label} are required.")
   }
-  columns <- dplyr::select(data, {{ columns }}) |> colnames()
-  if (rlang::is_empty(columns))
-    cli::cli_abort("The {.code columns} argument must select at least one column.")
   data_action <- rlang::enquo(data_action)
   report_listing <- rlang::enquo(report_listing)
   if (.is_quo_null(report_listing)) {
     report_listing <-
-      rlang::quo(select(., all_of(!!columns) & where(\(x) !inherits(x, !!class)))) |>
+      rlang::quo(select(., ends_with(".x") | ends_with(".y"))) |>
       structure(.Environment = rlang::caller_env()) # add the calling env as the quo env attribute
   }
 
-  # identify mis-matched columns -----------------------------------------------
+  # identify bad column names --------------------------------------------------
   bad_class_cols <-
-    select(data, all_of(columns) & where(\(x) !inherits(x, class))) |>
+    select(data, ends_with(".x") | ends_with(".y")) |>
     colnames()
 
   # construct condition quo ----------------------------------------------------
   quo_condition <-
     if (rlang::is_empty(bad_class_cols))
       condition <- rlang::quo(FALSE)
-    else
-      condition <- rlang::quo(FALSE)
+  else
+    condition <- rlang::quo(FALSE)
   condition <- condition |>
     structure(.Environment = rlang::caller_env()) # add the calling env as the quo env attribute
 
@@ -74,7 +69,7 @@ affirm_class <- function(data,
               id = id,
               priority = priority,
               data_frames = data_frames,
-              columns = paste(columns, collapse = ", "),
+              columns = "All Columns",
               report_listing = !!report_listing,
               data_action = !!data_action,
               error = error)